refactor(emotes-service): split util package into git and stack
Separate concerns: git operations stay in git package, stack
classification (prod/ephemeral) moves to stack package. Renames
GitCli to Cli now that package name provides context.
