+++ apps/emotes-service/infra/git/git_test.go
@@ -1,4 +1,4 @@
-package util
+package git
 
 import (
 	"strings"
